refactor: move stream history table to index
